+++ src/Ext/Anim/Body.cpp
@@ -456,6 +456,7 @@ void AnimExt::Serialize(T& Stm)
 		.Process(this->FiringAnim_LastFacing)
 		.Process(this->FiringAnim_LastCoords)
 		.Process(this->FirepowerMult)
+		.Process(this->AEDrawOffset)
 		;
 }