+++ src/Ext/AnimType/Body.cpp
@@ -80,6 +80,14 @@ void AnimTypeExt::LoadFromINIFile(CCINIClass* pINI)
 
 	this->Palette.LoadFromINI(pINI, pID, "CustomPalette");
 	this->XDrawOffset.Read(exINI, pID, "XDrawOffset");
+	this->XDrawOffset_ApplyBracketWidth.Read(exINI, pID, "XDrawOffset.ApplyBracketWidth");
+	this->XDrawOffset_InvertBracketShift.Read(exINI, pID, "XDrawOffset.InvertBracketShift");
+	this->XDrawOffset_BracketAdjust.Read(exINI, pID, "XDrawOffset.BracketAdjust");
+	this->XDrawOffset_BracketAdjust_Buildings.Read(exINI, pID, "XDrawOffset.BracketAdjust.Buildings");
+	this->YDrawOffset_ApplyBracketHeight.Read(exINI, pID, "YDrawOffset.ApplyBracketHeight");
+	this->YDrawOffset_InvertBracketShift.Read(exINI, pID, "YDrawOffset.InvertBracketShift");
+	this->YDrawOffset_BracketAdjust.Read(exINI, pID, "YDrawOffset.BracketAdjust");
+	this->YDrawOffset_BracketAdjust_Buildings.Read(exINI, pID, "YDrawOffset.BracketAdjust.Buildings");
 	this->HideIfNoOre_Threshold.Read(exINI, pID, "HideIfNoOre.Threshold");
 	this->Layer_UseObjectLayer.Read(exINI, pID, "Layer.UseObjectLayer");
 	this->AttachedAnimPosition.Read(exINI, pID, "AttachedAnimPosition");
@@ -144,6 +152,14 @@ void AnimTypeExt::Serialize(T& Stm)
 		.Process(this->Palette)
 		.Process(this->CreateUnitType)
 		.Process(this->XDrawOffset)
+		.Process(this->XDrawOffset_ApplyBracketWidth)
+		.Process(this->XDrawOffset_InvertBracketShift)
+		.Process(this->XDrawOffset_BracketAdjust)
+		.Process(this->XDrawOffset_BracketAdjust_Buildings)
+		.Process(this->YDrawOffset_ApplyBracketHeight)
+		.Process(this->YDrawOffset_InvertBracketShift)
+		.Process(this->YDrawOffset_BracketAdjust)
+		.Process(this->YDrawOffset_BracketAdjust_Buildings)
 		.Process(this->HideIfNoOre_Threshold)
 		.Process(this->Layer_UseObjectLayer)
 		.Process(this->AttachedAnimPosition)