+++ src/New/Entity/AttachEffectClass.cpp
@@ -112,7 +112,11 @@ AttachEffectClass::~AttachEffectClass()
 	if (it != AttachEffectClass::Array.end())
 		AttachEffectClass::Array.erase(it);
 
-	this->KillAnim();
+	if (this->Animation)
+	{
+		this->Animation->UnInit();
+		this->Animation = nullptr;
+	}
 
 	// the invoker may be mid-destruction with its extension already removed
 	if (this->Invoker)
@@ -325,13 +329,29 @@ void AttachEffectClass::AI_Temporal()
 
 void AttachEffectClass::AnimCheck()
 {
+	if (!this->Animation && this->CanShowAnim())
+		this->CreateAnim();
+}
+
+// Updates animation drawing logic that depends on state of other AE's on owner.
+// Called from TechnoExt::UpdateAEAnimDrawingLogic() on all AE's on owner if AE's are attached/detached/expire etc.
+// or if their animation state changes. Do not call directly.
+// Take care to to not invoke recursion - should not call any functions that call TechnoExt::UpdateAEAnimDrawingLogic() such as CreateAnim/KillAnim here.
+void AttachEffectClass::UpdateConditionalAnimDrawingLogic()
+{
 	if (this->Type->Animation_HideIfAttachedWith.size() > 0)
 	{
 		auto const pTechnoExt = TechnoExt::Fetch(this->Techno);
 
 		if (pTechnoExt->HasAttachedEffects(this->Type->Animation_HideIfAttachedWith, false, false, nullptr, nullptr, nullptr, nullptr))
 		{
-			this->KillAnim();
+			// Inlined because calling KillAnim() would cause recursive calls to this function.
+			if (this->Animation)
+			{
+				this->Animation->UnInit();
+				this->Animation = nullptr;
+			}
+
 			this->IsAnimHidden = true;
 			return;
 		}
@@ -339,8 +359,18 @@ void AttachEffectClass::AnimCheck()
 
 	this->IsAnimHidden = false;
 
-	if (!this->Animation && this->CanShowAnim())
-		this->CreateAnim();
+	if (this->Animation && this->Type->Animation_DrawOffsets.size() > 0)
+	{
+		auto const pAnimExt = AnimExt::Fetch(this->Animation);
+		auto const pTechnoExt = TechnoExt::Fetch(this->Techno);
+		pAnimExt->AEDrawOffset = Point2D::Empty;
+
+		for (auto const& drawOffset : this->Type->Animation_DrawOffsets)
+		{
+			if (drawOffset.RequiredTypes.size() < 1 || pTechnoExt->HasAttachedEffects(drawOffset.RequiredTypes, false, false, nullptr, nullptr, nullptr, nullptr, true))
+				pAnimExt->AEDrawOffset += drawOffset.Offset;
+		}
+	}
 }
 
 void AttachEffectClass::OnlineCheck()
@@ -455,6 +485,17 @@ void AttachEffectClass::CreateAnim()
 
 		pAnim->RemainingIterations = 0xFFu;
 		this->Animation = pAnim;
+		TechnoExt::Fetch(this->Techno)->UpdateAEAnimDrawingLogic();
+	}
+}
+
+void AttachEffectClass::KillAnim()
+{
+	if (this->Animation)
+	{
+		this->Animation->UnInit();
+		this->Animation = nullptr;
+		TechnoExt::Fetch(this->Techno)->UpdateAEAnimDrawingLogic();
 	}
 }
 
@@ -662,7 +703,7 @@ bool AttachEffectClass::ShouldBeDiscardedNow()
 			}
 		}
 	}
-	
+
 	if ((discardOn & DiscardCondition::Mission) != DiscardCondition::None)
 	{
 		auto const& missions = pTechno->Owner->IsControlledByHuman()
@@ -819,7 +860,7 @@ int AttachEffectClass::Attach(TechnoClass* pTarget, HouseClass* pInvokerHouse, T
 	{
 		pTargetExt->UpdateCumulativeAttachEffects(pType);
 	}
-	          
+
 	return attachedCount;
 }
 
@@ -927,6 +968,8 @@ AttachEffectClass* AttachEffectClass::CreateAndAttach(AttachEffectTypeClass* pTy
 	if (!currentTypeCount && cumulative && pType->CumulativeAnimations.size() > 0)
 		pAE->HasCumulativeAnim = true;
 
+	TechnoExt::Fetch(pTarget)->UpdateAEAnimDrawingLogic();
+
 	return pAE;
 }
 
@@ -1013,6 +1056,7 @@ int AttachEffectClass::DetachTypes(TechnoClass* pTarget, AEAttachInfoTypeClass c
 	if (detachedCount > 0)
 	{
 		const auto pExt = TechnoExt::Fetch(pTarget);
+		pExt->UpdateAEAnimDrawingLogic();
 
 		if (requiresRecalc)
 			pExt->RecalculateStatMultipliers();
@@ -1150,7 +1194,7 @@ void AttachEffectClass::TransferAttachedEffects(TechnoClass* pSource, TechnoClas
 
 			if (targetAttachEffect->GetType() == type)
 			{
-				currentTypeCount++;	
+				currentTypeCount++;
 
 				if (!cumulative)
 				{
@@ -1187,10 +1231,12 @@ void AttachEffectClass::TransferAttachedEffects(TechnoClass* pSource, TechnoClas
 
 		transferCount++;
 		it = pSourceExt->AttachedEffects.erase(it);
-	} 
+	}
 
 	if (transferCount > 0)
 	{
+		pTargetExt->UpdateAEAnimDrawingLogic();
+
 		if (requiresRecalc)
 			pTargetExt->RecalculateStatMultipliers();